DefaultOtions

+++ database/migrations/2025_05_15_002436_create_paises_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('codigo_iso_3', 3)->unique();
             $table->string('codigo_iso_2', 2)->unique();
             $table->string('nombre')->unique();
-            $table->unsignedTinyInteger('nivel')->nullable();
-            
+            $table->unsignedSmallInteger('nivel')->nullable();
+
         });
     }
